backend/requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
autopep8==1.7.0
pylint==2.14.5
pyjwt==2.4.0